+++ Anarchy/ChangeLog.md
@@ -2,4 +2,4 @@
 * Fixed tooltip related issues.
 * Updated localization. Added I18n Everywhere dependency.
 * Fixed Flaming chirper with disabled radio.
-* 
+* Added option to automatically disable anarchy toggle while brushing objects.